refactor: replace `recoil` to `jotai`
