+++ package.json
@@ -37,9 +37,11 @@
     "dayjs": "^1.11.13",
     "embla-carousel-react": "^8.6.0",
     "endpoint-client": "^0.1.1",
+    "fast-deep-equal": "^3.1.3",
     "history": "5.3.0",
     "i18next": "^24.2.3",
     "i18next-browser-languagedetector": "^8.0.4",
+    "jotai": "^2.12.2",
     "lucide-react": "^0.487.0",
     "mixpanel-browser": "^2.63.0",
     "next-themes": "^0.3.0",
@@ -51,7 +53,6 @@
     "react-i18next": "^15.4.1",
     "react-router": "^7.5.0",
     "react-slick": "0.30.3",
-    "recoil": "^0.7.7",
     "recoil-persist": "5.1.0",
     "slick-carousel": "^1.8.1",
     "sonner": "^1.7.4",
