+++ src/App.tsx
@@ -1,7 +1,6 @@
 import { BrowserRouter } from 'react-router';
 import { MainRouter } from './pages/router';
 import { Toaster } from './components/ui/sonner';
-import { RecoilRoot } from 'recoil';
 import { QueryClient, QueryCache, QueryClientProvider } from '@tanstack/react-query';
 import { APIResponseError } from 'endpoint-client';
 import './style.css';
@@ -29,13 +28,11 @@ const queryClient = new QueryClient({
 function App() {
   return (
     <QueryClientProvider client={queryClient}>
-      <RecoilRoot>
-        <BrowserRouter basename={import.meta.env.PUBLIC_URL}>
-          <ScrollToTop />
-          <MainRouter />
-          <Toaster />
-        </BrowserRouter>
-      </RecoilRoot>
+      <BrowserRouter basename={import.meta.env.PUBLIC_URL}>
+        <ScrollToTop />
+        <MainRouter />
+        <Toaster />
+      </BrowserRouter>
     </QueryClientProvider>
   );
 }
