+++ src/pages/layout/headerLayout.tsx
@@ -2,13 +2,12 @@ import { useEffect } from 'react';
 import { Outlet, useNavigate } from 'react-router';
 import { Header } from '@/containers/common/Header/Header';
 import { State } from '@/containers/common/Header/const/state';
-import { useRecoilValue, useSetRecoilState } from 'recoil';
-import { LoginState } from '@/recoil/atoms/atom';
+import { LoginState } from '@/atoms/atom';
 import { Footer } from '@/containers/common/Footer/Footer';
+import { useAtom } from 'jotai';
 
 export function Layout() {
-  const loginState = useRecoilValue(LoginState);
-  const setLoginState = useSetRecoilState(LoginState);
+  const [loginState, setLoginState] = useAtom(LoginState);
   const navigate = useNavigate();
 
   useEffect(() => {
