+++ src/pages/petition-notice/containers/PostSection/PetitionPostSection.tsx
@@ -9,17 +9,17 @@ import { PetitionSubcategoriesType } from '../../type';
 import { PetitionSubcategories } from '../../const';
 import { PetitionPostContent } from '@/components/deprecated/PetitionPostContent/PetitionPostContent';
 import { PetitionPostSectionSkeleton } from './PetitionPostSectionSkeleton';
-import { useRecoilValue } from 'recoil';
-import { SearchState } from '@/recoil/atoms/atom';
+import { SearchState } from '@/atoms/atom';
 import { useEffect } from 'react';
 import { useGetBoardPostSearch } from '@/hooks/api/get/useGetBoardPostSearch';
 import { useQueryClient } from '@tanstack/react-query';
+import { useAtom } from 'jotai';
 
 export function PetitionPostSection() {
   const navigate = useNavigate();
   const location = useLocation();
   const { currentPage, handlePageChange } = useCurrentPage(1);
-  const searchKeyword = useRecoilValue(SearchState);
+  const [searchKeyword] = useAtom(SearchState);
 
   const { selectedSubcategories, onSubcategorySelect } = useBoardSelect<PetitionSubcategoriesType>(
     PetitionSubcategories[0]
